+++ internal/repository/question/pg/converter/question.go
@@ -9,7 +9,7 @@ import (
 
 func QuestionFromPGSQL(questionRepo *modelRepo.Question) *model.Question {
 	questionServ := &model.Question{
-		ID:        int(questionRepo.ID),
+		ID:        int64(questionRepo.ID),
 		Text:      questionRepo.Text,
 		Type:      model.QuestionType(questionRepo.Type),
 		CreatedAt: questionRepo.CreatedAt,
