+++ internal/repository/questionset/pg/add_questions.go
@@ -8,7 +8,7 @@ import (
 	"github.com/Kosfedev/learn_go/internal/client/db"
 )
 
-func (r *repo) addQuestions(ctx context.Context, questionSetID int, questionIDs []int) error {
+func (r *repo) addQuestions(ctx context.Context, questionSetID int64, questionIDs []int64) error {
 	if len(questionIDs) == 0 {
 		return nil
 	}
