+++ internal/repository/questionset/pg/create.go
@@ -8,7 +8,7 @@ import (
 	"github.com/Kosfedev/learn_go/internal/repository/questionset/pg/converter"
 )
 
-func (r *repo) Create(ctx context.Context, newQuestionSet *model.NewQuestionSet) (int, error) {
+func (r *repo) Create(ctx context.Context, newQuestionSet *model.NewQuestionSet) (int64, error) {
 	newQuestionSetRepo := converter.NewQuestionSetToPGSQL(newQuestionSet)
 	query := db.Query{
 		Name:     "question_set_repository.create",
@@ -21,7 +21,7 @@ func (r *repo) Create(ctx context.Context, newQuestionSet *model.NewQuestionSet)
 		return 0, err
 	}
 
-	questionSetID := int(questionSetIDRepo)
+	questionSetID := int64(questionSetIDRepo)
 
 	// TODO: нужна транзакция
 	err = r.addQuestions(ctx, questionSetID, newQuestionSet.QuestionIDs)
