+++ internal/service/category/delete.go
@@ -4,7 +4,7 @@ import (
 	"context"
 )
 
-func (qcs *serv) Delete(ctx context.Context, id int) error {
+func (qcs *serv) Delete(ctx context.Context, id int64) error {
 	err := qcs.repo.Delete(ctx, id)
 	if err != nil {
 		return err
