+++ internal/service/question/add_options.go
@@ -6,7 +6,7 @@ import (
 	"github.com/Kosfedev/learn_go/internal/model"
 )
 
-func (qs *serv) AddOptions(ctx context.Context, questionID int, options []*model.NewQuestionOption) error {
+func (qs *serv) AddOptions(ctx context.Context, questionID int64, options []*model.NewQuestionOption) error {
 	err := qs.questionRepo.AddOptions(ctx, questionID, options)
 	if err != nil {
 		return err
