+++ internal/service/question/delete_options.go
@@ -4,7 +4,7 @@ import (
 	"context"
 )
 
-func (qs *serv) DeleteOptions(ctx context.Context, ids []int) error {
+func (qs *serv) DeleteOptions(ctx context.Context, ids []int64) error {
 	err := qs.questionRepo.DeleteOptions(ctx, ids)
 	if err != nil {
 		return err
