+++ internal/service/question/get.go
@@ -6,7 +6,7 @@ import (
 	"github.com/Kosfedev/learn_go/internal/model"
 )
 
-func (qs *serv) Get(ctx context.Context, questionID int) (*model.Question, error) {
+func (qs *serv) Get(ctx context.Context, questionID int64) (*model.Question, error) {
 	question, err := qs.questionRepo.Get(ctx, questionID)
 	if err != nil {
 		return nil, err
