+++ internal/service/question/remove_subcategories.go
@@ -4,7 +4,7 @@ import (
 	"context"
 )
 
-func (qs *serv) RemoveSubcategories(ctx context.Context, questionID int, subcategoryIDs []int) error {
+func (qs *serv) RemoveSubcategories(ctx context.Context, questionID int64, subcategoryIDs []int64) error {
 	err := qs.questionSubcategoryRepo.RemoveSubcategoriesFromQuestion(ctx, questionID, subcategoryIDs)
 	if err != nil {
 		return err
